refactor(P4b-1): extract Routing mixin verbatim from Executor
Per design R12 (refactor-under-green), inventory the shared @ivar state
first (see P4b-decomposition-embed.md §1) then move ~30 routing methods
into Inference::Executor::Routing without semantic edits.

Executor.rb: 2862 → 2402 lines (−460). routing.rb: 478 lines.
